Look Out (Arduino Codes and Packaging)
This file contains 3 different codes for 3 separate arduinos. The 1st code is an audio module for the chest. It basically detects a person's presence via HC-SR04 Ultrasonic sensor from an adjustable distance which then provides a soundtrack saying "Stay away" or "You will never get this treasure" which these soundtrack being randomized every time it triggers. The 2nd code is the chest locking mechanism with the help of a 5V Servo. It works by reading the input from the 3rd code on the portrait and connected via bluetooth. The servo will activate/unlock when a user manages to trigger the 3rd code. The box would open as long as the user covers the LDR on the 3rd module. Lastly the 3rd code is the going to be in the portrait. it contains 2 LDR sensors to get send information to the 2nd module via bluetooth. if a user covers the 2 LDRs longer than an adjustable period of time it will trigger the chest to open. shorter than that it would not send any input to the 2nd module.